requirements.txt
discord.py[voice]
discord-ext-voice-recv
httpx
openai
pyyaml
PyNaCl
gTTS
openai-whisper
torch
numpy